+++ src/RobotStudio.Scripting/GCodeWriter.cs
@@ -18,6 +18,7 @@ public static class GCodeWriter
     {
         HomeCommand => "G28",
         MoveToCommand move => WriteMove(move),
+        ScaraLinearMoveCommand move => WriteScaraMove(move),
         WaitCommand wait => $"G4 P{FormatNumber(wait.Duration.TotalMilliseconds)}",
         _ => throw new NotSupportedException($"G-code output does not support {command.GetType().Name}.")
     };
@@ -30,6 +31,17 @@ public static class GCodeWriter
             $"Z{FormatNumber(command.TargetPosition.Z)}";
 
         return command.RequestedVelocityMillimetersPerSecond is { } velocity
+            ? $"{result} F{FormatNumber(velocity * 60d)}"
+            : result;
+    }
+
+    private static string WriteScaraMove(ScaraLinearMoveCommand command)
+    {
+        var result =
+            $"G1 X{FormatNumber(command.TargetToolPose.X)} " +
+            $"Y{FormatNumber(command.TargetToolPose.Y)}";
+
+        return command.RequestedToolVelocityMillimetersPerSecond is { } velocity
             ? $"{result} F{FormatNumber(velocity * 60d)}"
             : result;
     }